refactor: move isArray into is-type
