+++ src/shared/hasMetaInfo.js
@@ -1,4 +1,4 @@
-import { isObject } from './typeof'
+import { isObject } from './is-type'
 
 // Vue $root instance has a _vueMeta object property, otherwise its a boolean true
 export default function hasMetaInfo(vm = this) {
