+++ src/shared/inMetaInfoBranch.js
@@ -1,4 +1,4 @@
-import { isUndefined } from './typeof'
+import { isUndefined } from './is-type'
 
 // a component is in a metaInfo branch when itself has meta info or one of its (grand-)children has
 export default function inMetaInfoBranch(vm = this) {
