--- src/shared/typeof.js
@@ -1,15 +1,0 @@
-export function isUndefined(arg) {
-  return typeof arg === 'undefined'
-}
-
-export function isObject(arg) {
-  return typeof arg === 'object'
-}
-
-export function isFunction(arg) {
-  return typeof arg === 'function'
-}
-
-export function isString(arg) {
-  return typeof arg === 'string'
-}
